requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
scipy>=1.7.0
stable-baselines3>=1.5.0
gym>=0.21.0
torch>=1.9.0
ta-lib>=0.4.24
pyyaml>=5.4.1
matplotlib>=3.4.3
seaborn>=0.11.2
tqdm>=4.62.3
optuna>=2.10.0
wandb>=0.12.0  # for experiment tracking
pytest>=6.2.5  # for testing
black>=21.9b0  # for code formatting
